# docs/05_API/OpenAPI.yaml
...
openapi: 3.1.0
info:
  title: MedRail API
  version: 0.1.0
  summary: Patient-consented health data layer under x402-paid intelligence endpoints, on Algorand.
  description: |-
    MedRail exposes **8 HTTP routes**: 3 priced with the x402 protocol (HTTP 402 +
    USDC settlement on Algorand) and 5 free. It also fronts an on-chain consent
    registry and audit log — the `MedRailConsent` smart contract, **App ID 768743428**
    on Algorand TestNet.

    ## What this API is not

    - **There is no database.** Durable state is Algorand global state, three box maps
      on App 768743428, and two static files compiled into the image.
    - **There is no LLM, model, or inference call.** The two "AI" endpoints are
      deterministic rule engines: 11 keyword rules with fixed weights, and a 14-row
      interaction table lookup. Every response carries a non-diagnostic disclaimer.
    - **There is no authentication.** No API keys, no tokens, no sessions, no accounts.
      Payment is the only gate, and it authenticates a *payment*, not a *caller*.
    - **There is no rate limiting** (requirement SEC-013 is NOT IMPLEMENTED).
    - **There is no idempotency.** No idempotency key, no deduplication, no replay
      window. A repeated paid call is a repeated payment.

    ## Middleware ordering — expect 402 before 400

    The x402 payment middleware is registered before every route handler
    (`api/src/app.ts:37-50`), so request-body validation never runs on an unpaid
...

...

...

...

...
  contact:
...
servers:
...
paths:
  /v1/triage:
    post:
      operationId: postTriage
      tags: [intelligence]
      summary: Rule-based clinical red-flag triage score ($0.02)
      description: |-
        Scores free-text symptoms against **11 hard-coded red-flag keyword groups** with
        fixed weights, sums the weights of all matched groups, caps the total at 100, and
        buckets it into an urgency band.

        **This is not a diagnosis and not a model.** The rule table is at
        `api/src/services/triageScorer.ts:32-44` and is fully inspectable. The function is
        pure: identical input always yields identical output (requirement NFR-009).

        **Weights:** mental health crisis 45 · possible stroke, FAST signs 40 · possible
        anaphylaxis 40 · possible cardiac chest pain 35 · respiratory distress 35 · loss of
        consciousness 30 · severe bleeding 30 · severe pain, unclear source 20 · high fever
        12 · persistent vomiting 10 · common mild symptom 2.

        **Bands:** `emergency` at score >= 60, `urgent` >= 30, `soon` >= 10, else `routine`.

        **Side effects: none.** No chain write, no disk write, no log line. The `symptoms`
        string is lower-cased in memory and discarded with the request.

        **Idempotency: none.** A retry is a second $0.02 payment.
        **Rate limit: none.**
      x-medrail-status: VALIDATED
      x-medrail-requirements: [FR-004, FR-005, FR-006, FR-009, AI-001, AI-002, AI-007, NFR-009]
      x-medrail-evidence: |-
        api/test/triageScorer.spec.ts (7 cases); api/test/x402-flow.spec.ts asserts a real
        402 with accepts[0].amount == "20000"; contracts/artifacts/e2e-proof.json records a
        real 200 obtained after settled transaction
        OYRQRKYA7WUKBVLWTOFJSJMZFBW7VCNGP5VGH5EBUJGRCVFQFJRQ.
      x-medrail-price:
        usd: "0.02"
        baseUnits: "20000"
        asset: USDC
        decimals: 6
      x-medrail-sideEffects: none
      x-medrail-idempotent: false
      x-medrail-rateLimited: false
      security: []
      parameters:
        - $ref: '#/components/parameters/PaymentSignature'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriageRequest'
            examples:
              emergency:
                summary: Cardiac plus respiratory red flags
                value:
                  symptoms: Sudden chest pain and shortness of breath
              routine:
                summary: A mild symptom only
                value:
                  symptoms: mild headache since yesterday
      responses:
        '200':
          $ref: '#/components/responses/TriageOk'
        '400':
          $ref: '#/components/responses/ValidationError'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/interaction-check:
    post:
      operationId: postInteractionCheck
      tags: [intelligence]
      summary: Check a medication list against known severe interaction pairs ($0.02)
      description: |-
        Compares a medication list against a curated table of **14 well-documented severe
        interaction pairs** shipped at `api/src/data/interactions.json` and read once,
        synchronously, at module load (`api/src/services/interactionChecker.ts:18`).

        Matching normalises each name with `trim().toLowerCase()` and then tests
        **bidirectional substring containment** in both directions
        (`m.includes(a) || a.includes(m)`), so partial names such as `"warfarin 5mg"` match.

        **Known weakness (requirement AI-006, NOT IMPLEMENTED):** the containment test is
        unanchored, so a 1-2 character medication name is contained by many table entries
        and produces false positives. The existing test for that input asserts only the
        disclaimer, so the behaviour is uncovered. Fix direction: token-boundary matching
        plus an RxNorm or synonym map.

        Severity distribution in the shipped table: 8 `major`, 3 `contraindicated`,
        3 `moderate`. `matches` is returned in **file order, not severity order**.

        **Side effects: none.** Medication names never leave the process.
        **Idempotency: none.** **Rate limit: none.**
      x-medrail-status: VALIDATED
      x-medrail-requirements: [FR-007, FR-008, FR-009, DATA-005, AI-004, AI-006, AI-007]
      x-medrail-evidence: api/test/interactionChecker.spec.ts (6 cases); api/test/x402-flow.spec.ts asserts a real 402.
      x-medrail-price:
        usd: "0.02"
        baseUnits: "20000"
        asset: USDC
        decimals: 6
      x-medrail-sideEffects: none
      x-medrail-idempotent: false
      x-medrail-rateLimited: false
      security: []
      parameters:
        - $ref: '#/components/parameters/PaymentSignature'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionRequest'
            examples:
              flagged:
                summary: A pair present in the reference table
                value:
                  medications: [warfarin, aspirin]
              clean:
                summary: No table pair matches
                value:
                  medications: [paracetamol, vitamin c]
      responses:
        '200':
          $ref: '#/components/responses/InteractionOk'
        '400':
          $ref: '#/components/responses/ValidationError'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/records/summary:
    post:
      operationId: postRecordsSummary
      tags: [records]
      summary: Consent-gated synthetic patient record summary ($0.05 + on-chain grant)
      description: |-
        Requires **both** a settled x402 payment **and** a currently-valid on-chain consent
        grant from `patientId` to `requesterAddress` for scope `records:summary` on
        `MedRailConsent` App ID 768743428.

        ## Behaviour

        1. The payment middleware settles $0.05 USDC.
        2. zod validates that both address fields are exactly 58 characters. **Length is
           the only validation — there is no checksum check** (requirement SEC-010,
           NOT IMPLEMENTED; see defect R-3).
        3. `check_access(patientId, requesterAddress, "records:summary")` is executed as a
           zero-fee `simulate()` call against algod. Nothing is submitted for this step.
        4. **Denied** — an audit entry with `action = "consent_denied"` is attempted, its
           failure is swallowed, and the API returns **403** with `paidButDenied: true`.
           The fee covers the on-chain verification regardless of outcome, the way a paid
           lookup API charges for a miss.
        5. **Allowed** — an audit entry with `action = "consent_checked"` is written to the
           chain as a real transaction signed by the operator/admin account, and the API
           returns 200 with `auditTxId` and `auditSequence`.

        ## Things an integrator must know

        - **`patientId` selects which grant is checked. It does NOT select data.** The
          `summary` object is a fixed constant (`api/src/routes/records.ts:15-21`) returned
          for every patient. There is no patient datastore (requirement DATA-004).
        - **`auditSequence` is a decimal STRING, not a JSON number.** The underlying value
          is a uint64/bigint outside JSON's safe-integer range and is `.toString()`-ed at
          `api/src/routes/records.ts:58`.
        - **`requesterAddress` is never authenticated — defect S-1 (CRITICAL).** Nothing
          binds it to the paying identity, and it is written verbatim into the immutable
          per-patient audit log, so a successful impersonation produces a permanent false
          attribution.
        - **Idempotency: none.** A retry after a timeout costs another $0.05 **and** writes
          a second audit entry.
        - **Defect R-2:** the success-path audit write is not wrapped in a catch, so a chain
          failure returns 500 after settlement. The caller pays and receives nothing.
        - **Cost per call:** permanently locks 58,900-59,300 microALGO of the application
          account's minimum balance, plus 18,900 microALGO on a patient's first-ever entry.
      x-medrail-status: UNVALIDATED
      x-medrail-requirements: [FR-010, FR-011, FR-012, FR-039, SEC-006, SEC-007, SEC-008, REL-002, DATA-004]
      x-medrail-evidence: |-
        No test exercises this handler. api/test/x402-flow.spec.ts asserts only the 402
        shape (accepts[0].amount == "50000"). total_audit_entries == 0 on App 768743428 and
        no audit box exists, so the success path has NEVER completed against the live
        contract (evidence gap E-1).
      x-medrail-defects: [S-1, R-1, R-2, R-3]
      x-medrail-price:
        usd: "0.05"
        baseUnits: "50000"
        asset: USDC
        decimals: 6
      x-medrail-sideEffects: |-
        Submits a real Algorand transaction (log_access, admin-signed) on BOTH the allowed
        and denied paths. Permanently increases the application account's minimum balance.
      x-medrail-idempotent: false
      x-medrail-rateLimited: false
      security: []
      parameters:
        - $ref: '#/components/parameters/PaymentSignature'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordsSummaryRequest'
            examples:
              liveAddresses:
                summary: The two accounts that appear in App 768743428's own grant history
                description: |-
                  Both live grant boxes currently hold status 2 (REVOKED), so this triple
                  returns 403 today.
                value:
                  patientId: S56WIB3XLUOXSUGAI6MDJ3GN3RV2NSE7RVV7DGNZJKDCJR4OJLLA6H7HEE
                  requesterAddress: CVYBERM3GTWGASZ7VUJLCBTVQA7AOPPTUW2Q3EISGR3XCRIKLM6SBIGUUA
      responses:
        '200':
          $ref: '#/components/responses/RecordsSummaryOk'
        '400':
          $ref: '#/components/responses/ValidationError'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '403':
          $ref: '#/components/responses/PaidButDenied'
        '500':
          $ref: '#/components/responses/InternalErrorAfterSettlement'
  /v1/consent/status:
    get:
      operationId: getConsentStatus
      tags: [consent]
      summary: Live on-chain consent validity for a (patient, requester, scope) triple (free)
      description: |-
        Free, unpaid, read-only. Executes the contract's `readonly` `check_access` method
        through `AtomicTransactionComposer.simulate()` — **zero fee, nothing submitted to
        the network** (requirement SEC-009).

        `granted` is `true` if and only if all three hold:

        1. the grant box exists;
        2. its `status` is 1 (`STATUS_GRANTED`);
        3. `expires_at == 0` (the "never expires" sentinel) **or**
           `Global.latest_timestamp < expires_at` (strict comparison, so a grant is already
           invalid *at* its expiry second).

        ## Two things that surprise integrators

        - **`scope` is free-form and is not normalised.** No trim, no case folding. The
          scope is hashed into the box key, so `"records:summary "` with a trailing space is
          a completely different grant and silently returns `granted: false`
          (requirement DATA-003).
        - **This free, unauthenticated endpoint requires the operator private key.** The
          simulated call needs a sender and a signer, so `OPERATOR_MNEMONIC` must be set
          (`api/src/services/algorand.ts:83-84`). Without it the endpoint returns 500.

        Two sequential algod round trips per request, with **no rate limit, no timeout, and
        no retry** — usable both to exhaust this service and to amplify traffic at AlgoNode
        (requirements SEC-013 and REL-003, both NOT IMPLEMENTED).
      x-medrail-status: UNVALIDATED
      x-medrail-requirements: [FR-013, SEC-009, SEC-010, SEC-013, DATA-003, REL-003, PERF-002]
      x-medrail-evidence: |-
        No automated test. Manually exercised by the reviewer: HTTP 200 in 505 ms, a single
        cold observation on a developer laptop. Not a p50, not an SLO; no latency budget
        exists.
      x-medrail-defects: [R-3]
      x-medrail-sideEffects: none
      x-medrail-idempotent: true
      x-medrail-rateLimited: false
      security: []
      parameters:
        - name: patient
          in: query
          required: true
          description: The patient's Algorand address. Exactly 58 characters; checksum NOT validated.
          schema:
            $ref: '#/components/schemas/AlgorandAddress'
          example: S56WIB3XLUOXSUGAI6MDJ3GN3RV2NSE7RVV7DGNZJKDCJR4OJLLA6H7HEE
        - name: requester
          in: query
          required: true
          description: The requester's Algorand address. Exactly 58 characters; checksum NOT validated.
          schema:
            $ref: '#/components/schemas/AlgorandAddress'
          example: CVYBERM3GTWGASZ7VUJLCBTVQA7AOPPTUW2Q3EISGR3XCRIKLM6SBIGUUA
        - name: scope
          in: query
          required: true
          description: |-
            Free-form consent scope. Minimum 1 character, no maximum, no normalisation.
            The API's own record route uses `records:summary`.
          schema:
            type: string
            minLength: 1
          example: 'records:summary'
      responses:
        '200':
          description: The live on-chain answer.
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/AccessControlAllowOrigin'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentStatusResponse'
              examples:
                revoked:
                  summary: The live triple on App 768743428 — box value decodes to status 2
                  value:
                    patient: S56WIB3XLUOXSUGAI6MDJ3GN3RV2NSE7RVV7DGNZJKDCJR4OJLLA6H7HEE
                    requester: CVYBERM3GTWGASZ7VUJLCBTVQA7AOPPTUW2Q3EISGR3XCRIKLM6SBIGUUA
                    scope: 'records:summary'
                    granted: false
        '400':
          description: |-
            One or more query parameters failed validation. Note the route-specific message.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
              example:
                error: 'invalid query — expected ?patient=&requester=&scope='
                details:
                  formErrors: []
                  fieldErrors:
                    patient: ['String must contain exactly 58 character(s)']
                    scope: ['String must contain at least 1 character(s)']
        '500':
          $ref: '#/components/responses/InternalError'
...
  /v1/consent/arc56:
    get:
      operationId: getConsentArc56
      tags: [consent]
      summary: The compiled ARC-56 application specification (free)
      description: |-
        Serves `contracts/artifacts/MedRailConsent.arc56.json` verbatim, so any third party
        can construct their own ABI calls against App 768743428 without this repository.

        Top-level keys: `name`, `structs` (`GrantRecord`, `AuditEntry`), `methods` (13),
        `arcs` (`[22, 28]`), `networks` (**an empty object — the deployed App ID is not
        recorded in the spec; use `/v1/consent/app-info`**), `state`, `bareActions`,
        `sourceInfo`.

        **Implementation note:** the file is `existsSync`-ed, `readFileSync`-read and
        `JSON.parse`-d on **every request** (`api/src/app.ts:63-68`) — a synchronous ~54 KB
        disk read and parse in the event loop, on a free, unauthenticated, unrate-limited
        route. Reading it once at module load would be strictly better. RECOMMENDED.
      x-medrail-status: UNVALIDATED
      x-medrail-requirements: [FR-015]
      x-medrail-evidence: No automated test. Verified by reading api/src/app.ts:63-69.
      x-medrail-sideEffects: none
      x-medrail-idempotent: true
      x-medrail-rateLimited: false
      security: []
      responses:
        '200':
          description: The ARC-56 application specification.
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/AccessControlAllowOrigin'
          content:
            application/json:
              schema:
                type: object
                description: |-
                  An ARC-56 Contract object. Not modelled field by field here; the served
                  document is authoritative.
                additionalProperties: true
                required: [name, structs, methods, arcs, state, bareActions]
                properties:
                  name:
                    type: string
                    const: MedRailConsent
                  arcs:
                    type: array
                    items:
                      type: integer
                    examples:
                      - [22, 28]
        '404':
          description: |-
            The compiled artifact is not present on disk at
            `__dirname/../../contracts/artifacts/MedRailConsent.arc56.json`. In a container
            this means the Dockerfile COPY step was skipped or the path resolution differs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                error: 'ARC-56 spec not found — has the contract been compiled?'
...
components:
  securitySchemes:
    x402Payment:
      type: apiKey
      in: header
      name: PAYMENT-SIGNATURE
      description: |-
        **Not an authentication scheme.** Modelled as `apiKey` only because OpenAPI has no
        native x402 type. The `PAYMENT-SIGNATURE` header carries a base64-encoded JSON
        `PaymentPayload` containing a signed Algorand transaction group, produced by an
        x402 v2 client such as `@x402/fetch` with `@x402/avm`'s `ExactAvmScheme`.

        It proves that **a payment settled**. It does **not** identify the caller to the
        route handler, and MedRail's handlers never decode it — which is precisely defect
        S-1. Every operation in this document therefore declares `security: []`: there is no
        authentication anywhere in this API.
  parameters:
    PaymentSignature:
      name: PAYMENT-SIGNATURE
      in: header
      required: false
      description: |-
        Base64-encoded JSON `PaymentPayload` — the signed x402 payment. Omit it to receive
        the 402 challenge in the `PAYMENT-REQUIRED` response header. Header lookup is
        case-insensitive.
      schema:
        type: string
        contentEncoding: base64
        contentMediaType: application/json
...
  schemas:
    AlgorandAddress:
      type: string
      minLength: 58
      maxLength: 58
      description: |-
        A base32 Algorand address.

        **Validation is length only** — the zod schema is `z.string().length(58)`. The
        checksum is NOT validated (requirement SEC-010, NOT IMPLEMENTED). A 58-character
        string that is not a valid address reaches `algosdk.decodeAddress` and throws,
        producing **HTTP 500** with the internal message echoed to the caller (defect R-3).
        The `pattern` below documents the true character set; the implementation does not
        enforce it.
      pattern: '^[A-Z2-7]{58}$'
      examples:
        - S56WIB3XLUOXSUGAI6MDJ3GN3RV2NSE7RVV7DGNZJKDCJR4OJLLA6H7HEE
        - CVYBERM3GTWGASZ7VUJLCBTVQA7AOPPTUW2Q3EISGR3XCRIKLM6SBIGUUA
    TriageRequest:
      type: object
      required: [symptoms]
      additionalProperties: true
      description: |-
        `additionalProperties: true` is accurate: the zod schema is a plain `z.object` with
        no `.strict()`, so unknown properties are ignored rather than rejected.
      properties:
        symptoms:
          type: string
          minLength: 1
          maxLength: 2000
          description: |-
            Free-text symptom description. **Never persisted** — lower-cased in memory,
            matched against the rule table, then discarded with the request
            (requirement AI-007).
...
    InteractionRequest:
      type: object
      required: [medications]
      additionalProperties: true
      properties:
        medications:
          type: array
          minItems: 2
          maxItems: 20
          description: |-
            Medication names. Each entry must be at least 1 character. Names are trimmed and
            lower-cased in memory and **never persisted** (requirement AI-007). There is no
            RxNorm or ATC normalisation and no synonym map.
          items:
            type: string
            minLength: 1
...
  responses:
    TriageOk:
      description: A triage score for the submitted symptom text.
      headers:
        PAYMENT-RESPONSE:
          $ref: '#/components/headers/PaymentResponse'
        Access-Control-Allow-Origin:
          $ref: '#/components/headers/AccessControlAllowOrigin'
        Access-Control-Expose-Headers:
          $ref: '#/components/headers/AccessControlExposeHeaders'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TriageResponse'
          examples:
            emergency:
              summary: Real output, reproduced from contracts/artifacts/e2e-proof.json
              value:
                score: 70
                band: emergency
                matchedFlags:
                  - possible cardiac chest pain
                  - respiratory distress
                disclaimer: >-
                  MedRail triage is a transparent keyword heuristic for hackathon
                  demonstration only. It is not a diagnosis, not a substitute for
                  professional medical judgment, and must never be the basis for a real care
                  decision. If this were real and urgent, call emergency services.
            routine:
              summary: A single low-weight rule matched
              value:
                score: 2
                band: routine
                matchedFlags: [common mild symptom]
                disclaimer: MedRail triage is a transparent keyword heuristic for hackathon demonstration only. ...
...
    PaymentRequired:
      description: |-
        No payment, an invalid payment, or a failed settlement.

        **The body is the empty object `{}`** — the entire payload is in the
        `PAYMENT-REQUIRED` header. Verified against a live capture.

        Also returned for a **malformed unpaid request**: the payment middleware runs before
        the route handler, so zod never sees an unpaid body. The repository's own test
        documents this (`api/test/x402-flow.spec.ts:48-59`).
      headers:
        PAYMENT-REQUIRED:
          $ref: '#/components/headers/PaymentRequired'
        Cache-Control:
          $ref: '#/components/headers/CacheControlNoStore'
        Access-Control-Allow-Origin:
          $ref: '#/components/headers/AccessControlAllowOrigin'
        Access-Control-Expose-Headers:
          $ref: '#/components/headers/AccessControlExposeHeaders'
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true
            description: Always `{}` in the current configuration.
          example: {}
    ValidationError:
      description: |-
        The request body failed zod validation. **Only reachable after a payment has
        settled**, because the payment middleware runs first (see the 402 response).
      headers:
        Access-Control-Allow-Origin:
          $ref: '#/components/headers/AccessControlAllowOrigin'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorBody'
          examples:
            triageMissingField:
              summary: 'POST /v1/triage with body {}'
              value:
                error: invalid request
                details:
                  formErrors: []
                  fieldErrors:
                    symptoms: [Required]
            interactionTooFew:
              summary: 'POST /v1/interaction-check with one medication'
              value:
                error: 'invalid request — provide at least 2 medications'
                details:
                  formErrors: []
                  fieldErrors:
                    medications: ['Array must contain at least 2 element(s)']
            recordsBadLength:
              summary: 'POST /v1/records/summary with a short address'
              value:
                error: invalid request
                details:
                  formErrors: []
                  fieldErrors:
                    patientId: ['String must contain exactly 58 character(s)']
    InternalError:
      description: |-
        An unhandled exception reached `app.onError` (`api/src/app.ts:58-61`), which logs it
        with `console.error` and returns **`err.message` verbatim** to the caller
        (requirement SEC-011, NOT IMPLEMENTED).

        Known causes, all reproduced or verified:

        - **R-1** — the facilitator is unreachable, so the 402 cannot be built:
          `"Failed to initialize: no supported payment kinds loaded from any facilitator."`
          Note the client gets **no `PAYMENT-REQUIRED` header, no `Retry-After`, and not a
          503** (requirement REL-001, NOT IMPLEMENTED).
        - **R-3** — a 58-character string that is not a valid address:
          `"wrong checksum for address"`, or `"Invalid base32 characters"` for a
          58-character string outside the base32 alphabet. A client input error surfaced as
          a server error.
        - **D-1** — `CONSENT_APP_ID` is unset and `contracts/artifacts/deploy_testnet.json`
          is absent, which is the default inside the container image:
          `"CONSENT_APP_ID is not set and contracts/artifacts/deploy_testnet.json was not found. Deploy the contract first (see docs/DEPLOYMENT.md)."`
        - `OPERATOR_MNEMONIC` unset: `"OPERATOR_MNEMONIC is not set — see docs/DEPLOYMENT.md"`.
        - **R-4** — algod unreachable or slow. `new algosdk.Algodv2("", server, "")` has no
          timeout and no retry (requirement REL-003, NOT IMPLEMENTED).

        Full catalogue: `./API_Error_Catalog.md`.
      headers:
        Access-Control-Allow-Origin:
          $ref: '#/components/headers/AccessControlAllowOrigin'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          examples:
            facilitatorDown:
              summary: R-1 — facilitator unreachable
              value:
                error: 'Failed to initialize: no supported payment kinds loaded from any facilitator.'
            badChecksum:
              summary: R-3 — 58 characters, invalid checksum
              value:
                error: wrong checksum for address
            appIdMissing:
              summary: D-1 — no App ID configured, the container default
              value:
                error: 'CONSENT_APP_ID is not set and contracts/artifacts/deploy_testnet.json was not found. Deploy the contract first (see docs/DEPLOYMENT.md).'
...